feat: signup api
net/http 회원가입 api 구현

go gin framework 적용 예정

+++ api/handler.go
@@ -1,6 +1,7 @@
 package api
 
 import (
+	"log"
 	"net/http"
 
 	"github.com/jeonjonghyeok/coinss-backend/psql"
@@ -8,14 +9,30 @@ import (
 	"github.com/jeonjonghyeok/coinss-backend/vo"
 )
 
+// CollectHost godoc
+// @Summary Host information collection.
+// @Description If it already exists, the changeable information is updated, and in the case of a new host, it is created and returned.
+// @Accept json
+// @Produce json
+// @Param apiver path string true "API version"
+// @Param project_id path string true "Project ID"
+// @Success 200 {object} vo.User
+// @Router /{apiver}/projects/{project_id}/hosts [post]
 func signup(w http.ResponseWriter, r *http.Request) {
 	var user vo.User
 	parseJSON(r.Body, &user)
+	exists, err := psql.IsExistUser(user.Email)
+	must(err)
+	log.Println(exists, user)
+	if exists {
+		panic(existUserError)
+	}
 
-	token, err := upbit.New(user.Accesskey, user.Secretkey)
+	id, err := psql.CreateUser(user)
 	must(err)
 
-	must(psql.CreateUser(user))
+	token, err := upbit.New(id, user.Accesskey, user.Secretkey)
+	must(err)
 
 	writeJSON(w, struct {
 		Token string `json:"token"`
